Compare modules' layer when tracking server CSS changes (#47927)
This PR fixes an issue reported by @feedthejim, that when we track CSS
module changes only the content hashes are compared. However the same
module can end up with totally different content in different layers,
hence this PR also adds the layer as part of the resource key so we
don't compare apples to oranges.

---------

Co-authored-by: JJ Kasper <jj@jjsweb.site>