Change 'post' to 'note'
This would make the documentation a lot easier to read at a glance, since we are not mixing the concept of a written 'Post' with the HTTP verb POST when documenting/discussing API implementation. 

May need further updating of the example projects to match the documentation, let me know if this is required and I will go ahead and update them.