Do not allow type Promise for props
A few type definitions were changed here #46046, including GetServerSidePropsResult. Within getServerSideProps, async functions have to be awaited; results are collected in the props object, which is then forwarded to the page. It does not make sense to return a Promise, which later has to be resolved into props within a page; it has to be an object identical to the props interface in GetStaticPropsResult. Close #48886