Correct some jest types (#48923)
# What?

This PR just refines some of the typescript types around `nextJest`, specifically refining the argument `customJestConfig` and the return value.

### Why?

Before this PR, the return type of `nextJest` was inferred by typescript as `any` which is inaccurate and would require type casting for typescript to enable it to be used in a jest configuration written in typescript.


I didn't add an issue for this since it was so trivial, but I can do so if it would be helpful.